spike: gets an answer for 2017, day 18, part 1

+++ src/main/java/fi/tuomax/adventofcode/commons/cpu/Argument.java
@@ -36,7 +36,7 @@ public class Argument
         }
     }
 
-    public Integer getValue() 
+    public Long getValue() 
     {
         return isRegister ? cpu.getRegister(register) : value;
     }
